requirements.txt
streamlit
ultralytics
opencv-python
pillow
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1